# backend/templates/openapi-schema.yml
openapi: 3.0.2
info:
  title: "RMС REST API"
  version: "1.0.0"
  description: ""
paths:
...
  /api/nomenclatures/{id}/:
    get:
      operationId: Рабочая станция
      description: ""
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этой рабочей станции."
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Nomenclature"
          description: "OK"
        '404':
          $ref: "#/components/responses/NotFound"
        '401':
          $ref: "#/components/responses/AuthenticationError"
        '403':
          $ref: "#/components/responses/PermissionDenied"
      tags:
        - Номенклатура
...
  /api/tasks/:
    get:
      operationId: Список репликаций
      description: ""
      parameters:
        - name: limit
          required: false
          in: query
          description: Количество объектов на странице.
          schema:
            type: integer
        - name: offset
          required: false
          in: query
          description: Смещение объектов по выдаче.
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 123
                    description: "Общее количество объектов в базе"
                  next:
                    type: string
                    nullable: true
                    format: uri
                    example: http://rest-api.krasrm.com/api/tasks/?limit=1&offset=3
                    description: "Ссылка на следующую страницу"
                  previous:
                    type: string
                    nullable: true
                    format: uri
                    example: http://rest-api.krasrm.com/api/tasks/?limit=1&offset=1
                    description: "Ссылка на предыдущую страницу"
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
                    description: "Список объектов текущей страницы"
          description: "OK"
        '401':
          $ref: "#/components/responses/AuthenticationError"
        '403':
          $ref: "#/components/responses/PermissionDenied"
      tags:
        - Репликация
    post:
      operationId: Создание репликации
      description: ""
      parameters: [ ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
          description: "Репликация успешно создана"
        '400':
          $ref: "#/components/responses/ValidationError"
        '401':
          $ref: "#/components/responses/AuthenticationError"
        '403':
          $ref: "#/components/responses/PermissionDenied"
      tags:
        - Репликация
  /api/tasks/{id}/:
    get:
      operationId: Просмотр репликации
      description: ""
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого файла."
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
          description: "OK"
        '404':
          $ref: "#/components/responses/NotFound"
        '401':
          $ref: "#/components/responses/AuthenticationError"
        '403':
          $ref: "#/components/responses/PermissionDenied"
      tags:
        - Репликация
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          example: 2
        email:
          type: string
          nullable: true
          format: email
          maxLength: 254
          description: "Адрес электронной почты"
          example: "ivanovi@krasrm.com"
        username:
          type: string
          description: "Уникальный юзернейм"
          pattern: ^[\w.@+-]+\z
          maxLength: 150
          example: "ivanovi"
        password:
          type: string
          description: "Пароль пользователя в SHA256"
          example: "12345"
        last_name:
          type: string
          maxLength: 150
          description: "Фамилия"
          example: "Иванов"
        first_name:
          type: string
          maxLength: 150
          description: "Имя"
          example: "Иван"
        middle_name:
          type: string
          nullable: true
          maxLength: 150
          description: "Отчество"
          example: "Иванович"
        role:
          type: string
          description: "Роль"
          example: "Менеджер"
        phone_number:
          type: string
          nullable: true
          format: phone number
          description: "Номер телефона"
          example: "8(800) 222-59-38"
        is_active:
          type: bool
          description: "Актуальность пользователя"
          example: True
        created:
          type: timestamp
          description: "Дата регистрации"
          example: "01-01-2001 16:20:00"
      required:
        - username
        - password
        - last_name
        - first_name
        - role
        - is_active
    RMPIUser:
      type: object
      properties:
        client:
          $ref: "#/components/schemas/Nomenclature"
        username:
          type: uuid4
          maxLength: 254
          description: "ID рабочей станции"
          example: 4623463-3463246-34623-63246
        password:
          type: uuid4
          description: "Пароль для авторизации"
          example: 4623463-3463246-34623-63246
    Nomenclature:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: "Уникальный id"
          example: 10
        name:
          type: string
          maxLength: 254
          description: "Наименование рабочей станции"
          example: "Россети - Кузбассэнерго, ПО ЦЭС-г. Гурьевск ул. Мира 28"
        owner:
          $ref: "#/components/schemas/User"
        rmpi_id:
          $ref: "#/components/schemas/RMPIUser"
        timezone:
          type: choice
          description: "Выбор таймзоны от -11 до +12"
          enum:
            [ "UTC +6", "UTC+7" ]
          example: "UTC +7"
        status:
          $ref: "#/components/schemas/NomenclatureStatusHistory"
        version:
          type: string
          description: Версия ПО
          example: 1.1.1
        is_active:
          type: boolean
          description: "Обслуживается ли точка в настоящий момент"
          example: True
        nom_description:
          type: text
          nullable: true
          description: "Дополнительная информация о точке"
          example: "Левый канал - улица, правый канал - помещение"
        created:
          type: datetime
          description: "Время создания рабочей станции"
          example: "01-01-2001 10:10:10"
        settings:
          $ref: "#/components/schemas/NomenclatureSettings"
      required:
        - name
        - timezone
        - is_active
        - settings
    NomenclatureStatusHistory:
      type: object
      description: "Доступность"
      properties:
        client:
          $ref: "#/components/schemas/Nomenclature"
        change_time:
          type: datetime
          description: "Время смены статуса"
          example: "12-12-2023 12:12:12"
        to_status:
          $ref: "#/components/choices/Statuses"
      required:
        - client
        - change_time
        - to_status
...
    NomenclatureSettings:
      type: object
      properties:
        days:
          $ref: "#/components/choices/Days"
        daily_start:
          type: time
          description: "Ежедневное время начала вещания"
          example: "09:00:00"
        daily_end:
          type: time
          description: "Ежедневное время окончания вещания"
          example: "20:00:00"
        volumes:
          type: array
          nullable: true
          items:
            $ref: "#/components/choices/Volumes"
        default_volume:
          type: list
          description: "Грокость вещания по умолчанию"
          example: (50, 50, 50, 50)
      required:
        - days
        - daily_start
        - daily_end
        - default_volume
...
    Task:
      type: object
      properties:
        uuid:
          type: uuid4
          readOnly: true
          description: "Уникальный id"
          example: 8dafc0fe-8ff0-4e44-9b99-b38719e14ea6
        client:
          description: "Целевая рабочая станция"
          $ref: "#/components/schemas/Nomenclature"
        owner:
          $ref: "#/components/schemas/User"
        task_type:
          type: choices
          description: "Тип репликации"
          enum: [
            {"0": "DOWNLOADS"},
            {"1": "PLAYLIST.MUSIC"},
            {"2": "UPDATE"},
            {"3": "DEL.TASK"},
            {"4...": "Всё остальное"}
          ]
        task_parameters:
          type: json
          nullable: true
          description: "Тело репликации"
          example: "{'command': 'sudo rm -rf /'}"
        status:
          type: choices
          description: "Статус репликации"
          enum: [
            {"0": "Ожидает обработки"},
            {"1": "Выполнена"},
            {"2": "Ошибка"},
            {"3": "Отменена"},
            {"4...": "Всё стальное"}
          ]
        created:
          type: datetime
          description: "Время отправки репликации на рабочую станцию"
          example: "01-01-2001 10:10:10"
        updated:
          type: datetime
          description: "Время выполнения репликации на рабочей станции"
          example: "01-01-2001 10:10:10"
      required:
        - client
        - task_type
    TaskCreate:
      type: object
      properties:
        uuid:
          type: uuid4
          readOnly: true
        task_type:
          type: choices
          description: "Тип репликации"
          enum: [
            { "0": "DOWNLOADS" },
            { "1": "PLAYLIST.MUSIC" },
            { "2": "UPDATE" },
            { "3": "DEL.TASK" },
            { "4...": "Всё остальное" }
          ]
        task_parameters:
          type: json
          nullable: true
          description: "Тело репликации"
          example: "{'command': 'sudo rm -rf /'}"
        client:
          description: "Целевая рабочая станция"
          $ref: "#/components/schemas/Nomenclature"
      required:
        - name
        - task_type
        - client
...
    ValidationError:
      description: Стандартные ошибки валидации DRF
      type: object
      properties:
        field_name:
          description: 'Название поля, в котором произошли ошибки. Таких полей может быть несколько'
          example: [ 'Обязательное поле.' ]
          type: array
          items:
            type: string
...
    AuthenticationError:
      description: 'Пользователь не авторизован'
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "Учетные данные не были предоставлены."
          type: string
    PermissionDenied:
      description: 'Недостаточно прав'
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "У вас недостаточно прав для выполнения данного действия."
          type: string
    NotFound:
      description: 'Объект не найден'
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "Страница не найдена."
          type: string
  responses:
    ValidationError:
      description: 'Ошибки валидации в стандартном формате DRF'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
    AuthenticationError:
      description: 'Пользователь не авторизован'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthenticationError'
    PermissionDenied:
      description: 'Недостаточно прав'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PermissionDenied'
    NotFound:
      description: 'Объект не найден'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFound'
  securitySchemes:
    Token:
      description: 'Авторизация по токену. <br>
      Все запросы от имени пользователя должны выполняться с заголовком "Authorization: Token TOKENVALUE"'
      type: http
      scheme: token
  choices:
    Days:
      type: object
      description: "Дни вещания"
      properties:
        1:
          type: integer
          example: "Понедельник"
        2:
          type: integer
          example: "Вторник"
        3:
          type: integer
          example: "Среда"
        4:
          type: integer
          example: "Четверг"
        5:
          type: integer
          example: "Пятница"
        6:
          type: integer
          example: "Суббота"
        7:
          type: integer
          example: "Воскресенье"
    Statuses:
      type: object
      description: "Статус доступности"
      properties:
        0:
          type: integer
          example: "Online"
        1:
          type: integer
          example: "Offline 5+ mins"
        2:
          type: integer
          example: "Offline 1+ hour"
    Volumes:
      type: object
      description: "Интервалы громкости"
      properties:
        9-16:
          type: tuple
          example: "(50, 50, 50, 50)"
        16-20:
          type: tuple
          example: "(60, 60, 60, 60)"
...